FPL countdown/requirements.txt
# HTTP requests and API handling
requests==2.32.0

# Data processing and manipulation
pandas==2.1.1
numpy==1.27.5

# Database connectivity for PostgreSQL
psycopg2-binary==2.9.7
SQLAlchemy==2.0.22

# Optional: for date/time handling
python-dateutil==2.8.2

# Optional: for environment variable management
python-dotenv==1.0.0
